non-mod-resources/skins/requirements.txt
Pillow>=10.0